apt-devel to BuildRequires; apt-libs to Requires
Using openscap to scan a system that use apt (such as Debian or Ubuntu) returns incomplete results. This warning is logged:
OVAL object 'dpkginfo_object' is not supported.

The problem is that openscap is built without apt support.

The fix, implemented in this commit, is to add apt as a build requirement.

Only Fedora provides the apt-devel package.
Therefore it is only included when building for Fedora.