Fix build on FreeBSD.
Now that FreeBSD has imported LLVM15, new errors are flagged.

Ideally the code that implicily converts integers should be fixed but
a first step is to use a compiler option that ignores the errors like
prior versions of LLVM and GCC do.

Additionaly, as in Linux, read_common_sizet() is also undefined in
FreeBSD.