Fix other occurences of oscap_htable_add
If the hash table owns its elements, its responsible for freeing
them. The callers of oscap_htable_add rely on the fact that the
elements will be later freed by oscap_htable_free. However, if
oscap_htable_add fails to insert the elements to the table,
then the caller needs to free them.