Prevent data race
Prevent data race on the static `indent` variable. The __oscap_dlprintf
function is called from multiple threads, which means multiple threads
can modify the `ident` static variable.

There is already a mutex (`__debuglog_mutex`) used in this code.
We only need to move the manipulation with the ident variable between
the lock and unlock operation on this mutex.