Fix issues when parsing profiles
When there is no white space between `<xccdf:select>` elements in
a profile in a benchmark, or in a custom profile in a tailoring,
every odd element was ignored because the XML node was skipped
by OpenSCAP when parsing.

This is caused by the situation that `xmlTextReaderRead` in
`oscap_to_start_element` in `xccdf_parse_remarks` jumps to the
next node but then `xmlTextReaderRead` at the end of the
`while loop` in `xccdf_profile_parse` jumps again to the next node.
If there is a white space node in between the element nodes, this
works fine, but if there is no node in between, an element gets
skipped.

The commit also adds some unit tests:
- `test_no_newline_between_select_elements` checks if no elements are
  skipped when parsing `<xccdf:select>` elements within `<xccdf:Profile>`
  elements
- `test_single_line_tailoring` checks if a tailoring file with no newlines
  between XML elements is processed correctly.

This is related to #1817.