Handle return value of probe_item_collect
The `probe_item_collect` function can fail and it can return various
return codes. But, the return codes aren't handled in the caller
function `process_file`. That means the `process_file` continues to
process items even if no other items can be collected due to memory
constraints or other problems. With this commit, we will read the return
code of `probe_item_collect` and react accordingly.