Make systemdunitproperty probe test more independent
The current test of the `systemdunitproperty` probe is relying on the
assumption that the `systemd-journald.service` unit doesn't have the
`Conflicts` property set. In Rawhide, this assumption is no longer true,
because `Conflicts` is now set to `soft-reboot.target` by default. This
change breaks our test. In this commit, we will make the test less
dependent on actual property value. Instead, we will check a dummy
property that is unlikely to occur.

Fixes: #2066