Improve counting collected items
The function SEXP_list_length turns out to be an unreliable
provider of information about the length of the list of the
collected items. It is unclear what this function returns.
Instead of using SEXP_list_length we will have a counter
variable that will count the amount of collected items
by the probe.

Fixes: #2076